Pull out API tests for multitext.
