+++ src/lib.rs
@@ -146,18 +146,11 @@ mod suffix_array;
 mod util;
 mod wrapper;
 
-// pub use crate::fm_index::FMIndex;
-// pub use crate::rlfmi::RLFMIndex;
-
-// pub use builder::SearchIndexBuilder;
 pub use character::Character;
 pub use frontend::{
-    FMIndex, FMIndexSearch, FMIndexSearchWithLocate, FMIndexWithLocate, RLFMIndex, RLFMIndexSearch,
-    RLFMIndexSearchWithLocate, RLFMIndexWithLocate, Search, SearchIndex, SearchIndexWithLocate,
-    SearchWithLocate,
+    FMIndex, FMIndexSearch, FMIndexSearchWithLocate, FMIndexWithLocate, MultiTextFMIndex,
+    MultiTextFMIndexSearch, MultiTextFMIndexSearchWithLocate, MultiTextFMIndexWithLocate,
+    RLFMIndex, RLFMIndexSearch, RLFMIndexSearchWithLocate, RLFMIndexWithLocate, Search,
+    SearchIndex, SearchIndexWithLocate, SearchWithLocate,
 };
 pub use iter::HeapSize;
-
-// pub use iter::{BackwardIterator, FMIndexBackend, ForwardIterator, HasPosition, HeapSize};
-// pub use multi_text::MultiTextFMIndex;
-// pub use search::Search;
